Add wokwi tests (#63)
Wokwi tests allow us to simulate the arduino in action, to catch
problems that may arise from running the code on hardware before the
code is approved. Currently, this simulates the temperature sensors, gps
sensor (also buzzer and GPS but no validation to see if it works, just
validation that process itself does not error out). It does this by
checking the `dataStr` and making sure it matches with what the data
output should be from these simulated sensors. This will need changing
if the `dataStr` changes or if sensors are added. It also keeps track of
pin location, allowing us to visualize pins at a glance (only of the
modules added but more to come). This could be used for future projects
and is useful to have in general. It will not be a required check as we
are limited to 50 minutes of pipeline runs a month and filtering this to
only run on changes to the `src` folder and wokwi files make it
impossible to require this pipeline, but code reviewers should not
approve PRs if this test fails without understanding the cause for the
failure.